@@ -33,7 +33,7 @@
         }
     ],
     "require": {
-        "php": "^7.4|^8.0",
+        "php": "^8.1",
         "ext-json": "*",
         "illuminate/auth": "^6|^7|^8.67|^9|^10",
         "illuminate/contracts": "^6|^7|^8.67|^9|^10",
